requirements.txt
python-telegram-bot==13.13
openai
requests
cryptography